worker, solver: update interfaces
