+++ solver/llbop/source.go
