+++ solver/signal.go
